feat: background header responsive
