--- src/components/headers/counter.ts
@@ -1,21 +1,0 @@
-const counter = (date: Date) => {
-    const countDownDate = date.getTime();
-    const now = new Date().getTime();
-
-    const distance = countDownDate - now;
-
-    const days = Math.floor(distance / (1000 * 60 * 60 * 24));
-    const hours = Math.floor((distance % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
-    const minutes = Math.floor((distance % (1000 * 60 * 60)) / (1000 * 60));
-    const seconds = Math.floor((distance % (1000 * 60)) / 1000);
-
-    return {
-        distance,
-        days,
-        hours,
-        minutes,
-        seconds
-    }
-}
-
-export { counter };
